requirements_dev.txt
responses==0.5.1
pytest==3.0.7
ipdb==0.10.1
ipython==6.0.0
factory-boy==2.8.1
